fix: use musl instead of GNU to avoid linking issue
The error in question was nix linking against libc.so instead of libc.so.6,
where the former is a linker script telling the linker to pick the latter, and
the latter is the actual library. I could not find out why the nix version was
linking against the wrong library.